requirements.txt
# ================================
# Agentic Threat Analyst — Okta Logs
# Requirements (LangChain v0.3+ compatible)
# ================================

# Core LLM & orchestration stack (split packages)
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Transformers + Accelerate (for Hugging Face pipelines)
transformers>=4.44.0
accelerate>=0.33.0

# PyTorch — optional on Apple Silicon (MPS works via system torch)
torch>=2.3.0; platform_system != "Darwin" or platform_machine != "arm64"

# Optional visualization / tensor utils
einops>=0.8.0

# YAML configuration and CLI tabulation
pyyaml>=6.0.1
tabulate>=0.9.0

# Optional: deterministic autogen simulation (conceptual only)
autogen>=0.3.0

# (macOS ARM users)
# If you encounter MPS backend issues, set:
#   export PYTORCH_ENABLE_MPS_FALLBACK=1
# before running.
